Переделал компоненты окон
